+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 ### Fixes
 - Improved error message when `image_name` matches an image that has a non-HDD type (e.g. CDROM) or is in a different location. The error now reports the found image's type and location.
 
+## 6.7.25
+### Features
+- File configuration failover for global resources: users, groups, target groups, s3keys, backupunits, contracts, object storage access keys
+### Fixes
+- Add debug log with error in case config file cannot be read due to whatever reason
+### Dependencies
+- Use shared v0.1.8
+
 ## 6.7.24
 ### Features
 - Added `location` field to several resources and data sources to support regional endpoints via file configuration. The field should only be used if a file configuration is provided (e.g. at `IONOS_CONFIG_FILE`).